# Allow linear and matmul to be consumed by nvFuser by default