--- a/notebooks/writing_a_trace_transform_cpu_offloading.ipynb
+++ b/notebooks/writing_a_trace_transform_cpu_offloading.ipynb
@@ -472,7 +472,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 13,
+   "execution_count": null,
    "metadata": {},
    "outputs": [
     {
@@ -507,7 +507,7 @@
     "model, args, kwargs = get_model_and_args()\n",
     "\n",
     "# Check against the vanilla `thunder.jit` model\n",
-    "expected = thunder.jit(model)(*args, **kwargs)\n",
+    "expected = thunder.jit(model, nv_enable_linear=False)(*args, **kwargs)\n",
     "\n",
     "grad_output = torch.randn_like(expected)\n",
     "expected_grads = torch.autograd.grad(expected, model.parameters(), grad_output)\n",
@@ -519,7 +519,7 @@
     "    expected_cpu = expected.to(\"cpu\")\n",
     "    expected_grads_cpu = tree_map(lambda t: t.to(\"cpu\"), expected_grads)\n",
     "\n",
-    "jmodel = thunder.jit(model, transforms=[CPUOffloading()])\n",
+    "jmodel = thunder.jit(model, nv_enable_linear=False, transforms=[CPUOffloading()])\n",
     "\n",
     "actual = jmodel(*args, **kwargs)\n",
     "\n",
@@ -642,7 +642,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 14,
+   "execution_count": null,
    "metadata": {},
    "outputs": [
     {
@@ -656,8 +656,8 @@
    "source": [
     "model, args, kwargs = get_model_and_args()\n",
     "\n",
-    "measurement_thunder = benchmark(thunder.jit(model), model, args, kwargs)\n",
-    "measurement_thunder_offload = benchmark(thunder.jit(model, transforms=[CPUOffloading()]), model, args, kwargs)\n",
+    "measurement_thunder = benchmark(thunder.jit(model, nv_enable_linear=False), model, args, kwargs)\n",
+    "measurement_thunder_offload = benchmark(thunder.jit(model, nv_enable_linear=False, transforms=[CPUOffloading()]), model, args, kwargs)\n",
     "\n",
     "del model, args, kwargs\n",
     "\n",