This change lowers peak memory usage of LitGPT implementations that use `mlp_class_name="GptNeoxMLP"` configuration (https://github.com/Lightning-AI/lightning-thunder/issues/1175, https://github.com/Lightning-AI/lightning-thunder/issues/1233, https://github.com/Lightning-AI/lightning-thunder/issues/246).

Updates from Jan 27th:
* `enable_linear=True` works as expected only together with allowing "shape only" fusions which were disabled in https://github.com/Lightning-AI/lightning-thunder/pull/1559, without reverting #1559 there's no change in peak used memory.
* `enable_linear=True` together with `enable_matmul=True` give substantial memory savings both for GptNeoxMLP- and LLaMAMLP-based models.

|           config (mlp_class_name)           |     Current main    | enable_linear + shape-only fusions | enable_linear + enable_matmul |
|:-------------------------------------------:|:---------------------:|:----------------------------------:|-------------------------------|
| stablecode-completion-alpha-3b (GptNeoxMLP) | 894.26 ms \| 76.99 GB | 894.84 ms \| 74.30 GB (1.036x)              | 919.44 ms \| 58.10 GB (1.325x)        |
| Llama-2-7b-hf (LLaMAMLP)                    | 342.67 ms \| 64.18 GB | 349.15 ms \| 63.11 GB (1.016x)              | 350.66 ms \| 57.55 GB (1.115x)        |

Old results from Oct 30th: 
| config (mlp_class_name)                     | Current main                | This PR               |
|---------------------------------------------|-----------------------|-----------------------|
| stablecode-completion-alpha-3b (GptNeoxMLP) | 889.68 ms \| 77.02 GB | 892.37 ms \| 74.33 GB |
| Llama-2-7b-hf (LLaMAMLP)                    | 336.06 ms \| 64.22 GB | 340.86 ms \| 64.18 GB |

Better memory usage comes from simplifying the setup for Thunder's fusion rematerialization. With this change, there are fewer "producer" fusions.

Commands for reproducing the results:
```
python thunder/benchmarks/benchmark_litgpt.py --model_name=stablecode-completion-alpha-3b --compile=dynamo_thunder
python thunder/benchmarks/benchmark_litgpt.py --model_name=Llama-2-7b-hf --compile=dynamo_thunder
```

cc @Priya2698, @wujingyue