@Priya2698 2024-10-30
We did not run Thunder benchmarks using nvfuser linear. Should we run other benchmarks as well before enabling it by default?

Additionally, @wujingyue needed to remove support for 1D weights to facilitate DID-aware execution. We might have to add an additional check on Thunder side or use `unsqueeze-squeeze` operators to support it. I believe we currently do not have cases exercising 1D weights in Thunder so this should not break anything right now.