@IvanYashchuk 2024-10-31
> are there `enable_foobar`s other than linear and matmul?

Yes, there's also `nv_enable_sdpa`. I didn't enable matmul in this PR because it leads to excessive memory usage in my tests. I haven't tried `nv_enable_sdpa` I thin we want to continue using the one from cuDNN for performance.