@wujingyue 2024-11-16
CI with torch-nightly is now passing with https://github.com/NVIDIA/Fuser/issues/3369 fixed. CI with older versions of torch (and therefore older versions of nvFuser) still fail, because we can fix a past version. @IvanYashchuk, can you up nvFuser's version and enable linear only for that?