@wujingyue 2024-11-18
I'm unsure about the lightning-thunder (ipynb) test. Is that an infra failure or indeed a regression?