@IvanYashchuk 2024-11-25
For some reason, there was a timeout in the cell execution. Let's try rerunning the failed job.