@IvanYashchuk 2024-11-27
@kiya00, could you please check what's going on with the Notebooks CI job?