@kevinstephano 2024-12-03
> @kiya00 Is there a bug for the slowness issue? @kevinstephano, do you think it's worth investigating?
While there is a real issue, between 20 to 30 matmuls in a fusion that leads to an explosion in segmentation time, I am not sure how often we are going to see so many matmuls in one fusion.  It would be interesting to see if this comes up in actual models.