@IvanYashchuk 2025-01-17
> I think it would be good to re-gather the memory impact (baseline and with PR) before merging. We should see much better memory footprint in the baseline, potentially rendering this PR moot.

Sure, I will do that.