@IvanYashchuk 2025-01-27
> I think it would be good to re-gather the memory impact (baseline and with PR) before merging. We should see much better memory footprint in the baseline, potentially rendering this PR moot.

The baseline numbers didn't change much from October (77.02 GB -> 76.99 GB, 64.22 GB -> 64.18 GB). 

`enable_linear=True` doesn't give any memory or performance changes by itself, it requires reverting changes from https://github.com/Lightning-AI/lightning-thunder/pull/1559, but the good news is that using `enable_linear=True` and `enable_matmul=True` together give great results now (it wasn't the case in Oct).

Before merging this change, I want to check the impact on all models we track.