@IvanYashchuk 2025-02-18
There is no benefit in throughput (memory savings are not enough to increase the batch size for better throughput) for a larger set of models so not pursuing further as the default. In addition, we can't rely on sending linear and matmul to nvFuser when using the TransformerEngine executor.